src/test/java/lib/persistent/PersistentTreeMapTest.java
/* Copyright (C) 2016  Intel Corporation
 *
 * This program is free software; you can redistribute it and/or
 * modify it under the terms of the GNU General Public License
 * version 2 only, as published by the Free Software Foundation.
 * This file has been designated as subject to the "Classpath"
 * exception as provided in the LICENSE file that accompanied
 * this code.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License version 2 for more details (a copy
 * is included in the LICENSE file that accompanied this code).
 *
 * You should have received a copy of the GNU General Public License
 * version 2 along with this program; if not, write to the Free
 * Software Foundation, Inc., 51 Franklin Street, Fifth Floor,
 * Boston, MA  02110-1301, USA.
 */

package lib.persistent;

import java.util.*;

public class PersistentTreeMapTest {

    public static void main(String[] args) {
        System.out.println("****************PersistentTreeMap Tests****************");
        testInsertion();
        testRemoval();
        testIteration();
        testSubMap();
        testPutAll();
        testNullValue();
        testSubmapOthers();
    }

    public static void testInsertion() {
        System.out.println("****************Testing insertion****************");

        PersistentTreeMap<PersistentByteBuffer, PersistentByteBuffer> map = TestUtil.getOrCreateMap("m1");
        map.clear();
        assert(map.size() == 0);

        PersistentByteBuffer key = TestUtil.createByteBuffer(5, "hello");
        PersistentByteBuffer val = TestUtil.createByteBuffer(5, "world");

        assert(map.get(key) == null);

        PersistentByteBuffer out = map.put(key, val);
        assert(out == null);
        assert(map.get(key) != null);
        assert(TestUtil.shouldEqual(map.get(key), "world"));
        assert(map.size() == 1);

        val.rewind().put("java".getBytes()).rewind();
        assert(map.size() == 1);
        assert(map.get(key) != null);
        assert(TestUtil.shouldEqual(map.get(key), "javad"));

        key.rewind().put("hey".getBytes()).rewind();
        assert(map.size() == 1);
        assert(map.get(key) != null);
        assert(TestUtil.shouldEqual(map.get(key), "javad"));

        PersistentByteBuffer key2 = TestUtil.createByteBuffer(5, "heylo");
        assert(TestUtil.shouldEqual(map.get(key2), "javad"));

        PersistentByteBuffer key3 = TestUtil.createByteBuffer(5, "hello");
        assert(map.get(key3) == null);

        map.clear();
    }

    public static void testRemoval() {
        System.out.println("****************Testing removal****************");

        PersistentTreeMap<PersistentByteBuffer, PersistentByteBuffer> map = TestUtil.getOrCreateMap("m1");
        assert(map.size() == 0);

        PersistentByteBuffer key = TestUtil.createByteBuffer(5, "hello");
        PersistentByteBuffer val = TestUtil.createByteBuffer(5, "world");

        map.put(key, val);
        assert(map.size() == 1);

        PersistentByteBuffer key2 = TestUtil.createByteBuffer(5, "hello");
        PersistentByteBuffer out = map.remove(key2);
        assert(TestUtil.shouldEqual(out, "world"));
        assert(map.size() == 0);
        assert(map.isEmpty() == true);

        map.put(key2, val);
        assert(TestUtil.shouldEqual(map.get(key), "world"));
        assert(map.size() == 1);

        out = map.remove(key);
        assert(TestUtil.shouldEqual(out, "world"));
        assert(map.size() == 0);
        assert(map.isEmpty() == true);

        map.clear();
    }

    public static void testIteration() {
        System.out.println("****************Testing iteration****************");

        PersistentTreeMap<PersistentByteBuffer, PersistentByteBuffer> map = TestUtil.getOrCreateMap("m2");
        map.clear();
        assert(map.size() == 0);

        TreeMap<PersistentByteBuffer, PersistentByteBuffer> tmap = new TreeMap<PersistentByteBuffer, PersistentByteBuffer>();

        PersistentByteBuffer key1 = TestUtil.createByteBuffer(5, "hello");
        PersistentByteBuffer val1 = TestUtil.createByteBuffer(5, "world");

        PersistentByteBuffer key2 = TestUtil.createByteBuffer(5, "hi");
        PersistentByteBuffer val2 = TestUtil.createByteBuffer(5, "java");

        PersistentByteBuffer key3 = TestUtil.createByteBuffer(5, "good");
        PersistentByteBuffer val3 = TestUtil.createByteBuffer(5, "bye");

        PersistentByteBuffer key4 = TestUtil.createByteBuffer(5, "hillo");
        PersistentByteBuffer val4 = TestUtil.createByteBuffer(5, "world");

        map.put(key1, val1);
        map.put(key2, val2);
        map.put(key3, val3);
        map.put(key4, val4);

        assert(map.size() == 4);

        tmap.put(key1, val1);
        tmap.put(key2, val2);
        tmap.put(key3, val3);
        tmap.put(key4, val4);

        Set<Map.Entry<PersistentByteBuffer, PersistentByteBuffer>> s1 = map.entrySet();
        Set<Map.Entry<PersistentByteBuffer, PersistentByteBuffer>> s2 = tmap.entrySet();
        assert(s1.size() == s2.size());
        assert(s1.size() == 4);

        Iterator<Map.Entry<PersistentByteBuffer, PersistentByteBuffer>> sit1 = s1.iterator();
        Iterator<Map.Entry<PersistentByteBuffer, PersistentByteBuffer>> sit2 = s2.iterator();

        while (sit1.hasNext() || sit2.hasNext()) {
            Map.Entry<PersistentByteBuffer, PersistentByteBuffer> e1 = sit1.next();
            Map.Entry<PersistentByteBuffer, PersistentByteBuffer> e2 = sit2.next();
            assert(e1.getKey().equals(e2.getKey()));
            assert(e1.getValue().equals(e2.getValue()));
        }

        Set<PersistentByteBuffer> ks1 = map.keySet();
        Set<PersistentByteBuffer> ks2 = tmap.keySet();
        assert(ks1.size() == ks2.size());
        assert(ks1.size() == 4);

        Iterator<PersistentByteBuffer> ksit1 = ks1.iterator();
        Iterator<PersistentByteBuffer> ksit2 = ks2.iterator();

        while (ksit1.hasNext() || ksit2.hasNext()) {
            assert(ksit1.next().equals(ksit2.next()));
        }

        Collection<PersistentByteBuffer> vc1 = map.values();
        Collection<PersistentByteBuffer> vc2 = tmap.values();
        assert(vc1.size() == vc2.size());
        assert(vc1.size() == 4);

        Iterator<PersistentByteBuffer> vit1 = vc1.iterator();
        Iterator<PersistentByteBuffer> vit2 = vc2.iterator();

        while (vit1.hasNext() || vit2.hasNext()) {
            assert(vit1.next().equals(vit2.next()));
        }

        PersistentByteBuffer key5 = TestUtil.createByteBuffer(5, "good");
        PersistentByteBuffer val5 = TestUtil.createByteBuffer(5, "bye");

        assert(map.containsKey(key5) == true);
        assert(map.containsValue(val5) == true);

        PersistentByteBuffer key6 = TestUtil.createByteBuffer(5, "goo");
        PersistentByteBuffer val6 = TestUtil.createByteBuffer(5, "weird");

        assert(map.containsKey(key6) == false);
        assert(map.containsValue(val6) == false);

        assert(map.firstKey().equals(tmap.firstKey()));
        assert(map.get(map.firstKey()).equals(tmap.get(tmap.firstKey())));

        assert(map.lastKey().equals(tmap.lastKey()));
        assert(map.get(map.lastKey()).equals(tmap.get(tmap.lastKey())));

        map.put(key6, val6);
        tmap.put(key6, val6);
        assert(map.size() == 5);
        assert(s1.size() == s2.size());
        assert(s1.size() == 5);

        sit1 = s1.iterator();
        sit2 = s2.iterator();

        while (sit1.hasNext() || sit2.hasNext()) {
            Map.Entry<PersistentByteBuffer, PersistentByteBuffer> e1 = sit1.next();
            Map.Entry<PersistentByteBuffer, PersistentByteBuffer> e2 = sit2.next();
            assert(e1.getKey().equals(e2.getKey()));
            assert(e1.getValue().equals(e2.getValue()));
        }

        map.clear();
    }

    public static void testSubMap() {
        System.out.println("****************Testing submap****************");
        PersistentTreeMap<PersistentByteBuffer, PersistentByteBuffer> map = TestUtil.getOrCreateMap("m3");
        TreeMap<PersistentByteBuffer, PersistentByteBuffer> staging = new TreeMap<PersistentByteBuffer, PersistentByteBuffer>();

        assert(map.size() == 0);

        PersistentByteBuffer keyA = TestUtil.createByteBuffer(1, "a");
        PersistentByteBuffer keyE = TestUtil.createByteBuffer(1, "e");
        PersistentByteBuffer keyI = TestUtil.createByteBuffer(1, "i");
        PersistentByteBuffer keyO = TestUtil.createByteBuffer(1, "o");
        PersistentByteBuffer keyU = TestUtil.createByteBuffer(1, "u");

        PersistentByteBuffer val = TestUtil.createByteBuffer(3, "val");

        staging.put(keyA, val);
        staging.put(keyE, val);
        staging.put(keyI, val);
        staging.put(keyO, val);
        staging.put(keyU, val);

        map.putAll(staging);

        System.out.println("****************Testing headmap****************");
        headMapTests(map, val);

        map.clear();
        map.putAll(staging);

        System.out.println("****************Testing tailmap****************");
        tailMapTests(map, val);

        map.clear();
        map.putAll(staging);

        System.out.println("****************Testing submap****************");
        subMapTests(map, val);

        map.clear();
    }

    /*
     * All submap tests break down into four big cases, based on where a provided key can fall (example of headmap)
     * - Before beginning of map (smaller than all existing keys)    - case 1
     * - Between two existing keys on map                            - case 2
     * - Exactly equal to one of the existing keys on map            - case 3
     * - After end of map (larger than all existing keys)            - case 4
     * Similar cases for tailmap (larger than all existing keys for case 1, etc.) and submap.
     */

    private static void headMapTests(PersistentTreeMap map, PersistentByteBuffer val) {

        PersistentByteBuffer headKey = TestUtil.createByteBuffer(1, "Z");
        SortedMap<PersistentByteBuffer, PersistentByteBuffer> hm = map.headMap(headKey);   // Z < a, case 1
        assert(hm.size() == 0);

        PersistentByteBuffer key1 = TestUtil.createByteBuffer(1, "U");
        map.put(key1, val);
        assert(hm.size() == 1);
        assert(TestUtil.shouldEqual(hm.firstKey(), "U"));

        headKey = TestUtil.createByteBuffer(1, "f");
        hm = map.headMap(headKey);   // e < f < i, case 2
        assert(hm.size() == 3);  // E, a, e

        headKey = TestUtil.createByteBuffer(1, "o");
        hm = map.headMap(headKey);   // case 3
        assert(hm.size() == 4);  // E, a, e, i

        headKey = TestUtil.createByteBuffer(1, "w");
        hm = map.headMap(headKey);   // w > u, case 4
        assert(hm.size() == 6);  // E, a, e, i, o, u

        PersistentByteBuffer key2 = TestUtil.createByteBuffer(1, "v");
        hm.put(key2, val);
        assert(hm.size() == 7);
        assert(map.size() == 7);

        PersistentByteBuffer key3 = TestUtil.createByteBuffer(1, "z");
        map.put(key3, val);
        assert(hm.size() == 7);
        assert(map.size() == 8);

        PersistentByteBuffer key4 = TestUtil.createByteBuffer(1, "w");
        map.put(key4, val);
        assert(hm.size() == 7);
        assert(map.size() == 9);

        boolean caught = false;
        try {
            hm.put(key3, val);
        } catch (IllegalArgumentException e) {
            caught = true;
        }
        assert(caught);

        headKey = TestUtil.createByteBuffer(1, "O");
        hm = map.headMap(headKey);

        PersistentByteBuffer tailKey = TestUtil.createByteBuffer(1, "E");
        SortedMap<PersistentByteBuffer, PersistentByteBuffer> htm = hm.tailMap(tailKey);
        assert(htm.size() == 0);

        PersistentByteBuffer key5 = TestUtil.createByteBuffer(1, "L");
        map.put(key5, val);
        assert(htm.size() == 1);
        assert(hm.size() == 1);
        assert(map.size() == 10);

        PersistentByteBuffer startKey = TestUtil.createByteBuffer(1, "F");
        PersistentByteBuffer endKey = TestUtil.createByteBuffer(1, "L");
        SortedMap<PersistentByteBuffer, PersistentByteBuffer> htsm = htm.subMap(startKey, endKey);
        assert(htsm.size() == 0);

        PersistentByteBuffer key6 = TestUtil.createByteBuffer(1, "G");
        hm.put(key6, val);
        assert(htsm.size() == 1);
        assert(hm.size() == 2);
        assert(htm.size() == 2);
        assert(map.size() == 11);

        PersistentByteBuffer headKey2 = TestUtil.createByteBuffer(1, "P");
        caught = false;
        try {
            SortedMap<PersistentByteBuffer, PersistentByteBuffer> hhm = hm.headMap(headKey2);
        } catch (IllegalArgumentException e) {
            caught = true;
        }
        assert(caught);
    }

    private static void tailMapTests(PersistentTreeMap map, PersistentByteBuffer val) {

        PersistentByteBuffer tailKey = TestUtil.createByteBuffer(1, "w");
        SortedMap<PersistentByteBuffer, PersistentByteBuffer> tm = map.tailMap(tailKey);  // w > u, case 1
        assert(tm.size() == 0);

        PersistentByteBuffer key1 = TestUtil.createByteBuffer(1, "z");
        map.put(key1, val);
        assert(tm.size() == 1);

        tailKey = TestUtil.createByteBuffer(1, "p");
        tm = map.tailMap(tailKey);    // o < p < u, case 2
        assert(tm.size() == 2);  // u, z

        tailKey = TestUtil.createByteBuffer(1, "i");
        tm = map.tailMap(tailKey);    // case 3
        assert(tm.size() == 4); // i, o, u, z

        tailKey = TestUtil.createByteBuffer(1, "W");
        tm = map.tailMap(tailKey);    // W > a, case 4
        assert(tm.size() == 6);

        PersistentByteBuffer key2 = TestUtil.createByteBuffer(1, "Z");
        tm.put(key2, val);
        assert(tm.size() == 7);
        assert(map.size() == 7);

        PersistentByteBuffer key3 = TestUtil.createByteBuffer(1, "L");
        map.put(key3, val);
        assert(tm.size() == 7);
        assert(map.size() == 8);

        PersistentByteBuffer key4 = TestUtil.createByteBuffer(1, "W");
        map.put(key4, val);
        assert(tm.size() == 8);
        assert(map.size() == 9);

        PersistentByteBuffer key5 = TestUtil.createByteBuffer(1, "0");
        boolean caught = false;
        try {
            tm.put(key5, val);
        } catch (IllegalArgumentException e) {
            caught = true;
        }
        assert(caught);

        tailKey = TestUtil.createByteBuffer(1, "u");
        tm = map.tailMap(tailKey);
        assert(tm.size() == 2);

        PersistentByteBuffer headKey = TestUtil.createByteBuffer(1, "z");
        SortedMap<PersistentByteBuffer, PersistentByteBuffer> thm = tm.headMap(headKey);
        assert(thm.size() == 1);

        PersistentByteBuffer key6 = TestUtil.createByteBuffer(1, "w");
        map.put(key6, val);
        assert(thm.size() == 2);
        assert(tm.size() == 3);
        assert(map.size() == 10);

        PersistentByteBuffer startKey = TestUtil.createByteBuffer(1, "u");
        PersistentByteBuffer endKey = TestUtil.createByteBuffer(1, "w");
        SortedMap<PersistentByteBuffer, PersistentByteBuffer> thsm = thm.subMap(startKey, endKey);
        assert(thsm.size() == 1);

        assert(thsm.remove(endKey) == null);
        assert(map.get(endKey) != null);

        PersistentByteBuffer tailKey2 = TestUtil.createByteBuffer(1, "d");
        caught = false;
        try {
            SortedMap<PersistentByteBuffer, PersistentByteBuffer> ttm = tm.tailMap(tailKey2);
        } catch (IllegalArgumentException e) {
            caught = true;
        }
        assert(caught);

        PersistentByteBuffer tailKey3 = TestUtil.createByteBuffer(1, "u");
        PersistentByteBuffer tailKey4 = TestUtil.createByteBuffer(1, "w");
        SortedMap<PersistentByteBuffer, PersistentByteBuffer> ttm = tm.tailMap(tailKey3);
        SortedMap<PersistentByteBuffer, PersistentByteBuffer> tttm = ttm.tailMap(tailKey4);

        assert(ttm.size() == 3);
        assert(tttm.size() == 2);
    }

    private static void subMapTests(PersistentTreeMap map, PersistentByteBuffer val) {

        PersistentByteBuffer startKey1 = TestUtil.createByteBuffer(1, "f");
        PersistentByteBuffer endKey1 = TestUtil.createByteBuffer(1, "z");
        SortedMap<PersistentByteBuffer, PersistentByteBuffer> sm = map.subMap(startKey1, endKey1);
        assert(sm.size() == 3);

        startKey1 = TestUtil.createByteBuffer(1, "Z");
        endKey1 = TestUtil.createByteBuffer(1, "z");
        sm = map.subMap(startKey1, endKey1);
        assert(sm.size() == 5);

        startKey1 = TestUtil.createByteBuffer(1, "a");
        endKey1 = TestUtil.createByteBuffer(1, "u");
        sm = map.subMap(startKey1, endKey1);
        assert(sm.size() == 4);

        startKey1 = TestUtil.createByteBuffer(1, "G");
        endKey1 = TestUtil.createByteBuffer(1, "a");
        sm = map.subMap(startKey1, endKey1);
        assert(sm.size() == 0);

        startKey1 = TestUtil.createByteBuffer(1, "u");
        endKey1 = TestUtil.createByteBuffer(1, "z");
        sm = map.subMap(startKey1, endKey1);
        assert(sm.size() == 1);

        startKey1 = TestUtil.createByteBuffer(1, "f");
        endKey1 = TestUtil.createByteBuffer(1, "h");
        sm = map.subMap(startKey1, endKey1);
        assert(sm.size() == 0);

        PersistentByteBuffer key1 = TestUtil.createByteBuffer(1, "f");
        map.put(key1, val);
        assert(sm.size() == 1);
        assert(map.size() == 6);

        PersistentByteBuffer key2 = TestUtil.createByteBuffer(1, "h");
        boolean caught = false;
        try {
            sm.put(key2, val);
        } catch (IllegalArgumentException e) {
            caught = true;
        }
        assert(caught);

        startKey1 = TestUtil.createByteBuffer(1, "U");
        endKey1 = TestUtil.createByteBuffer(1, "d");
        sm = map.subMap(startKey1, endKey1);
        assert(sm.size() == 1);

        PersistentByteBuffer key3 = TestUtil.createByteBuffer(1, "Z");
        sm.put(key3, val);
        assert(sm.size() == 2);
        assert(map.size() == 7);

        PersistentByteBuffer headKey = TestUtil.createByteBuffer(1, "b");
        SortedMap<PersistentByteBuffer, PersistentByteBuffer> shm = sm.headMap(headKey);
        assert(shm.size() == 2);

        PersistentByteBuffer key4 = TestUtil.createByteBuffer(1, "c");
        sm.put(key4, val);
        assert(sm.size() == 3);
        assert(shm.size() == 2);
        assert(map.size() == 8);

        PersistentByteBuffer tailKey = TestUtil.createByteBuffer(1, "a");
        SortedMap<PersistentByteBuffer, PersistentByteBuffer> shtm = shm.tailMap(tailKey);
        assert(shtm.size() == 1);

        PersistentByteBuffer key5 = TestUtil.createByteBuffer(1, "W");
        shm.put(key5, val);
        assert(shm.size() == 3);
        assert(sm.size() == 4);
        assert(shtm.size() == 1);
        assert(map.size() == 9);
    }

    public static void testPutAll() {
        System.out.println("****************Testing putAll****************");
        PersistentTreeMap<PersistentByteBuffer, PersistentByteBuffer> map = TestUtil.getOrCreateMap("m4");

        assert(map.size() == 0);

        TreeMap<PersistentByteBuffer, PersistentByteBuffer> staging = new TreeMap<PersistentByteBuffer, PersistentByteBuffer>();

        PersistentByteBuffer key1 = TestUtil.createByteBuffer(5, "reese");
        PersistentByteBuffer val1 = TestUtil.createByteBuffer(5, "cups");

        PersistentByteBuffer key2 = TestUtil.createByteBuffer(5, "hi");
        PersistentByteBuffer val2 = TestUtil.createByteBuffer(5, "java");

        PersistentByteBuffer key3 = TestUtil.createByteBuffer(5, "good");
        PersistentByteBuffer val3 = TestUtil.createByteBuffer(5, "bye");

        PersistentByteBuffer key4 = TestUtil.createByteBuffer(5, "hillo");
        PersistentByteBuffer val4 = TestUtil.createByteBuffer(5, "world");

        staging.put(key1, val1);
        staging.put(key2, val2);
        staging.put(key3, val3);
        staging.put(key4, val4);

        map.putAll(staging);
        assert(map.equals(staging));
        assert(staging.equals(map));

        map.clear();
    }

    public static void testNullValue() {
        System.out.println("****************Testing null value****************");

        PersistentTreeMap<PersistentByteBuffer, PersistentByteBuffer> map = TestUtil.getOrCreateMap("m5");
        map.clear();
        assert(map.size() == 0);

        PersistentByteBuffer key = TestUtil.createByteBuffer(5, "hello");
        map.put(key, null);

        assert(map.get(key) == null);
        assert(map.size() == 1);
    }

    public static void testSubmapOthers() {
        System.out.println("****************Testing Submap Other Functions****************");

        PersistentTreeMap<PersistentByteBuffer, PersistentByteBuffer> map = TestUtil.getOrCreateMap("m4");

        assert(map.size() == 0);

        PersistentByteBuffer key1 = TestUtil.createByteBuffer(5, "reese");
        PersistentByteBuffer val1 = TestUtil.createByteBuffer(5, "cups");

        PersistentByteBuffer key2 = TestUtil.createByteBuffer(5, "hi");
        PersistentByteBuffer val2 = TestUtil.createByteBuffer(5, "java");

        PersistentByteBuffer key3 = TestUtil.createByteBuffer(5, "goody");
        PersistentByteBuffer val3 = TestUtil.createByteBuffer(5, "byeby");

        PersistentByteBuffer key4 = TestUtil.createByteBuffer(5, "hillo");
        PersistentByteBuffer val4 = TestUtil.createByteBuffer(5, "world");

        map.put(key1, val1);
        map.put(key2, val2);
        map.put(key3, val3);
        map.put(key4, val4);

        PersistentByteBuffer startKey = TestUtil.createByteBuffer(5, "hi");
        PersistentByteBuffer endKey = TestUtil.createByteBuffer(5, "world");
        SortedMap<PersistentByteBuffer, PersistentByteBuffer> sm = map.subMap(startKey, endKey);

        TreeMap<PersistentByteBuffer, PersistentByteBuffer> staging = new TreeMap<PersistentByteBuffer, PersistentByteBuffer>();

        PersistentByteBuffer key5 = TestUtil.createByteBuffer(5, "goo");
        PersistentByteBuffer val5 = TestUtil.createByteBuffer(5, "weird");

        PersistentByteBuffer key6 = TestUtil.createByteBuffer(5, "word");
        PersistentByteBuffer val6 = TestUtil.createByteBuffer(5, "excel");

        staging.put(key5, val5);
        staging.put(key6, val6);

        boolean caught = false;
        try {
            sm.putAll(staging);
        } catch (IllegalArgumentException e) {
            caught = true;
        }
        assert(caught);

        staging.clear();

        sm.clear();
        assert(map.size() == 1);
        assert(TestUtil.shouldEqual(map.firstKey(), "goody"));
        assert(TestUtil.shouldEqual(map.get(map.firstKey()), "byeby"));

        PersistentByteBuffer hihi = TestUtil.createByteBuffer(5, "hihi");
        caught = false;
        try {
            sm.compute(key3, (k, v) -> (v == null ? hihi : (v.rewind().put("java".getBytes()).rewind())));
        } catch (IllegalArgumentException e) {
            caught = true;
        }
        assert(caught);
        assert(TestUtil.shouldEqual(map.get(key3), "byeby"));

        caught = false;
        try {
            sm.compute(key3, (k, v) -> (v.rewind().put("java".getBytes()).rewind()));
        } catch (NullPointerException e) {
            caught = true;
        }
        assert(caught);
        map.clear();
    }
}
